refactor: Update import paths for DokanLogo component
